point manager test: refactor

+++ src/main/java/ar/edu/unq/desapp/grupod/argconectadabackend/model/Donor.java
@@ -5,25 +5,25 @@ import java.util.Map;
 
 public class Donor extends User {
 	
-	private int points;
+	private Double points;
 	
 	private RewardProgram rewardProgram;
 	
-	private Map<String, Integer> pointsRegistry = new HashMap<String, Integer>();
+	private Map<String, Double> pointsRegistry = new HashMap<String, Double>();
 	
 	private String nickName;
 
 	public Donor(String name, String password, String email, String nickName, RewardProgram rewardProgram) {
 		super(name, password, email);
-		this.points = 0;
+		this.points = 0.0;
 		this.nickName = nickName;
 	}
 
-	public int getPoints() {
+	public Double getPoints() {
 		return points;
 	}
 
-	public void setPoints(int points) {
+	public void setPoints(Double points) {
 		this.points = points;
 	}
 
@@ -39,7 +39,7 @@ public class Donor extends User {
 		this.points += points;
 	}
 	
-	public void addPointsToRegister(String nameOfProject, int points) {
+	public void addPointsToRegister(String nameOfProject, Double points) {
 		this.pointsRegistry.put(nameOfProject, points);
 	}
